[chore]: PR리뷰용 주석 설정 (#65)
