+++ app/src/main/java/com/texthip/thip/data/repository/BookRepository.kt
@@ -1,8 +1,8 @@
-package com.texthip.thip.data.model.repository
+package com.texthip.thip.data.repository
 
 import com.texthip.thip.data.model.base.handleBaseResponse
 import com.texthip.thip.data.model.book.response.BookSavedResponse
-import com.texthip.thip.data.model.service.BookService
+import com.texthip.thip.data.service.BookService
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -11,7 +11,7 @@ class BookRepository @Inject constructor(
     private val bookService: BookService
 ) {
 
-    // 저장된/모임 책 조회 API 연동
+    /** 저장된 책 또는 모임 책 목록 조회 */
     suspend fun getBooks(type: String): Result<List<BookSavedResponse>> {
         return try {
             bookService.getBooks(type)
