+++ app/src/main/java/com/texthip/thip/data/service/BookService.kt
@@ -1,4 +1,4 @@
-package com.texthip.thip.data.model.service
+package com.texthip.thip.data.service
 
 import com.texthip.thip.data.model.base.BaseResponse
 import com.texthip.thip.data.model.book.response.BookListResponse
@@ -7,6 +7,7 @@ import retrofit2.http.Query
 
 interface BookService {
 
+    /** 저장된 책 또는 모임 책 목록 조회 */
     @GET("books")
     suspend fun getBooks(
         @Query("type") type: String
